+++ CHANGELOG.md
@@ -8,7 +8,9 @@ No date
 
 26.01.2023
 
-Bugfix.
+- `--cache-downloads` option. In 28.0, `mist` cached downloads into `/private/tmp/com.ninxsoft.mist`. This is now optional.
+
+Bugfixes:
 - `mist` result is now correctly interpreted when checking for a newer version.
 - The `--update` option now triggers an invalid installer to be overwritten.